refactor: onboarding components and restructure files
