+++ apps/app/src/app/(onboarding)/onboarding/layout.tsx
@@ -1,0 +1,11 @@
+interface OnboardingLayoutProps {
+  readonly children: React.ReactNode;
+}
+
+export default function OnboardingLayout({ children }: OnboardingLayoutProps) {
+  return (
+    <main className="flex h-svh w-full flex-col items-center justify-center">
+      {children}
+    </main>
+  );
+}
