+++ apps/app/src/components/_onboarding/onboarding-stepper.tsx
