+++ apps/app/src/components/_onboarding/workspace-setup.tsx
