+++ apps/app/src/components/onboarding/choose-style.tsx
@@ -1,0 +1,9 @@
+'use client';
+
+export function ChooseStyle() {
+  return (
+    <div>
+      <h1>Choose Style Component</h1>
+    </div>
+  );
+}
