+++ apps/app/src/components/onboarding/invite.tsx
@@ -1,0 +1,9 @@
+'use client';
+
+export function Invite() {
+  return (
+    <div>
+      <h1>Invite Component</h1>
+    </div>
+  );
+}
