+++ apps/app/src/components/onboarding/ready.tsx
@@ -1,0 +1,9 @@
+'use client';
+
+export function Ready() {
+  return (
+    <div>
+      <h1>Ready Component</h1>
+    </div>
+  );
+}
